Got debugging on our effects, found the source of the trouble- need to map the data down more, in each effect, right now we are passing misc data all the way down to the point we modify firestore db, which explains why update and delete arent working (id we use to find a specific document is either undefined or otherwise inaccessible)
in the case of update and create, we are adding
extra data to the db. So it looks like for each effect
we either need to map down an extra level, or
change the way data is passed in to action or when
the action is dispatched from the component.
either way, we are finally getting there with
NGRX + firestore CRUD!