Got game feature selector / state working, we can now view games in the database, IFF they have an id property. Havent tested out adding, deletion, updating yet, but this will require we refactor game-info for the admin-game-form to work.
Really just refactored game feature because I want
to do a proof of concept for a better organized
way of importing and exporting from the store.
this is was started the whole refactoring escapade
to begin with- we needed unique action names to
export them from actions directory. however, now
that im trying to experiment with getting that
working, im realizing its not as practical as i initially
imagined it. going to keep on pushing for this,
but at the time, only one feature at a time works
for admin.

What wed like to do is write an AdminStoreIndex class
which exports all the actions, storeModule.forFeature
and EffectsModule.forFeature calls we need for the admin
module, so that setting up the store in the admin module
is as simple as importing that class to the admin.module imports.