code is bug free and servable, project is not.
Our custom game-info selector that grabs
game info in state with a certain family, is not working now.

I'm afraid we arent out of the tunnel with reorganization
troubles, there is possible circular imports
between game info selector and reducer
, and Im uncertain of where entity related code should
live. Lot of debugging to do to get this small feature to work
but im on it. again, want to optimize the feature chain
(actions, reducer, selectors, effects) on a few small features
before I expand ngrx coverage to the whole site.
We are picking up steam...