Basically got our custom selector working by nerfing the use of entity in gameInfo reducer, so we rewrote it as a regular reducer without the entityAdapter. The selector was getting squirelly because state is defined differently and selectors must handle it diffferently when dealing with entity
considering if we want to ditch entity all together'
needs further consideration, but i dont see any
discernable benefit to keeping it. It causes a ton of bugs
and requires more convoluted code. On the other hand,
seen other ngrx developers swear by it. conisder the pros
and cons (make a list!) before you nerf it from the whole
app