Very stuck still on the custom selector for getting a family type from gameInfo state.
At this point, I have posted my code to stackoverflow
because this bug is very puzzeling and mysterious to me
Furthermore, after solving this, I'm hoping this will be
the last data driven problem, so we can turn our
attention towards testing, layout, styling, applying
bootstrap and angular material.

We got a solution to the last error for getFamily
customSelector. Problem is that now, it silently fails,
and I am unsure whether this is a problem with
the selector, how I use it, or how I am thinking of
the problem all together. Either way, I will be glad
when we push through here.