implemented full crud interface for admin-content. Finally at the point where I am very much glad we upgraded to NGRX. This turned out such a mess when we were working directly with firebase through a service.
Im sure its not bug free, but at first glance I see no problems.

Worked on documentation a little, proofread, organized
 and added a link from readme to contributing guide.

created a firebase service for getting server timestamp,
registered to the app.module, as well as a services
directory for it to live in. perhaps consider moving
it to core module, seems to be an appropriate place
for that to live.

created a directory for documentation to live, this will
require further attention on a documentation day.