Tried to debug a smoll problem with content crud, which is that when editing an existing piece of content, the existing tags arent loaded by multi-select-chip component. Wasnt able to solve at this time, so we add it to the bug list
Added some organization to documentation folder,
we now have readme, buglog, changelog and contributing
plus a folder for uml, split into class and sequence diagrams.

formated the Contributing file, added an index,
the links arent working and Im not sure what the problem
is. added this to buglog, its a problem but Im not going
to stop my roll today to take care of it, Tags are important
to viewing and organizing content though, so I making
a note of it, so it gets resolved in time.

Lets try to make a practice out of using and editing this
set of documentation frequently:
- readme
- buglog
- changelog
- contributing
To be used to greatest effect, They should reflect
a very uptodate picture of the site.