Got update function working pretty well. to resolve the bug, we needed to make two changes, one essential, the other is quality of life improvement
1) component is initialized with formControls for
title and description set to empty string,
in order to make both piececes of data assigned
to the form control when we wanted to update,
we need to reassign both those to empty strings

2) need to check if the form is pristine, when updating,
and disable the submit button when true. This
keeps users from dispatching the update action,
when no update has been made.