Attempted making game c.r.u.d services by dialog component, but realized this gets super messy when you involve observables.
not saying this shouldnt be done, but lets put it
off. its a nice to have feature, lets focus on the site core
and come back later, just opened up a whole can of worms
trying to get a prototype of it working.

did some minor refactoing of game interface, moved
it into the models folder.