All things in the site which used activatedRoute have now been updated to use ngrx-router-store.
Came across a bug in both admin-game-form
and admin-forum-form. That prevents them from
being updated. I plan on next making these forms
reactively with form builder, not to say tdf forms
are to blame, but generally I think the alternative
works better when you are editing existing objects.
plus reactivity works well with observable datastreams
like that from the store.

as it is, these components are the only ones which rely
on subscription to get at the underlying database values.
everything else leverages the async pipe within the template,
which is wicked cool, ME THINKS