Still hacking away at admin-game-form
created a 'connectForm' directive which
allows us to provide async data to a form.

most irritating this... It was working fine, even
printing update object data to console and template,
and then we tried to add support for select> categories
creators, platforms, and the whole thing stopped working
with update data. I have no idea what I did but this
bothers me and again really need to learn debugging
strategies, need to do it and stop just saying it.