Made admin-forum-form reactive and did a bit of debugging. Meanwhile, trying to add accessiblitity to these forms with additional bootstrap classes and svg icons. I also dont think the URL validator is working. because reactive forms needs validators not template attributes.
Havent finished debugging either admin- thing - forms,
the chief problem is that the data isnt loaded when the
component is displayed, likelybecause the observable isnt
ready to subscribe or the db call hasnt finished executing.
it probably a good idea to add some app.component level
css + js that says :
```
while !observable.omitted
    styles = 'loading.spinner.active'
```

Also seems to be an issue with that route param selector..
again... it works reliably yet bugiliy in admin-game-form and
admin-forum-form, does not work at all in forum or game
module when you try to view an object. Ive made
enancements to game-view component to respond to this
problem, but it should be the next item on your todo list.

also for consideration- should admin users really be able
to add , delete and modify forums with ease ? I was thining
because i made a link in game-view for user to report problem,
but if an admin deletes the 'help and support' forum,
this link becomes a deadlink