Identified two problems mentioned in last commit 1. route params were not working because the feature module routing modules were not updated to account for the fact that the name of the route param were changed from :id to :<contentType>Id 2. svgs icons need to be in a div / parent container with a bootstrap contextual class
added a Not found display for both games and forum view
which displays if the route param doesnt look up an existing
game. Id like to spend today getting these more accessible
as well as the corresponding routes, maybe make them
fed by a query param with the game or forum name, so then
we could go to bug reporting like 'forum/help/new' instead
of a messy uuid as the 2nd parameter.

Also would be cool to get the chexkbox filters on game
display working, get consoles crud working in admin-game
info, be able to add games to watchlists, news, streams etc,
able to add and work with threads via admin-forum-form,
(but first, consider the note in my last commit!!!)