Hacking away at console form, but also realized we made a design error in the process of replicating the set up for game info form.
we should have one form at the top , or as a routed
component, which manages both game info and console
crud. the way we have it set up now is redundant and
counter intutitive *see todo comment/note in admin
-consoles.

in this commit,
-got admin-console to print out list
of consoles per maker
-created dropdown form in admin-console
-admin-console emits create, update and delete
events, which are received by parent
-logged the data in parent, confirming
that the form returns the correct data that
the action dispatch requires.
-create Console works, but these actions are
commented out so as to not fill the db up
with poorly formatted, invalid or test data