In the process of major refactoring for the admin module, making more components reusable throughout the module (dumb components).
admin-info-table :
displays an observable data source in a table
and emits create, update and delete events

admin-info-table-manager:
shows a list of choices for which type of
data to display in table, contains both
table and form for editing

admin-dashboard-form:
facilitates adminstration of table data
by listening for create, update events,
populates existing data when it becomes
available.

game-info/console-selected.service:
a custom event listener/ behaivor subject,
prevents us from having to manually bubeling
up existing data from table to form, instead
the form subscribes to this subject.

currently trying to make zippy open using
input property, so that when a user wants to
create a new game info or console, the zippy
opens. something is going on behind the scenes,
with the lifecycle hooks i think, which is messing
with this. We can probably skip this, if it isnt
easily resolved,