Ignoring the event handling / store connection side of things to work on straightening out details with the components and their archieture and communication.
got console tab working and console-form
populated when a user wants to edit a console.

have a weird problem with the form data, not
sure what the cause is but looks like it could
use some debugging/ breakpoint analysis,
seems that all our selects dont work in creation
mode for the form, only when we are editing
and the data has been populated, can you make
a choice. Tried moving the variable initialization
into ngOnInit, but perhaps ngOnChanges makes
better sense, since at least one of the select values
is provided by Input.