Did some troubleshooting of the admin-game-info and admin-consoles interfaces, these need some touching up but largely are error free upon performing known use cases.
started implementing the same interfaces for
admin-forum-info and the new component
admin-threads. mostly just focused on forum-info
for now. in the process, realized that admin-dashboard-form
isnt as reusable as we had hoped, two major issues.
one work around would be to add conditionals so
that it supports crud of forum-info and threads as
well as game-info and consoles. There is a lesson
deep in here, that if you try making complicated
components infinitly reusable, they are far too abstract
to be useful in production. reusable dumb components
need to be barebones simple, cant have too many moving
parts.

also did some "work" with styling the navbar, we had
planned to save styling for last things to work on,
but uncommented the sass for the navbar. it is tricky
to debug, while it looks good and is responsive, its tricky
to fine tune.

Im thinking this is just a for-now solution, since we are
reusing this navbar sass from a react project, its
a huge stylesheet with hundreds of lines, now we
are in angular and can break it down into manageable
chunks, then it will be easier to parse how the styles
effect different sub components, and make changes,
as of now, it sort of feels like troubleshooting this navbar
is taking shots in the dark, unless its something obvious
like changing the color or fontsize.