In the process of implementing thread reducer, action, effects, selector, implemented all the remaining reducers, actions required for the site (at least what i think is still required): -thread -rating -review -suggestion -feedback -user
made a index file tree for models, so we can
import them on mass in reducer index file easily.

didnt finish implementing all these, didnt get to the
selectors, but it realized it may be a good idea to
make the names of selectors more specific. ie
`getGameInfoFamily` instead of `getFamily`
and also make the selectors more uniform, so that
you can add a filter to game-info, forum-info and
consoles and threads.