modified the name of selectors. they all now use select instead of get and contain the name of the object from store they select modified and deugged to adapt to these.
Didnt get around to finishing up the selector
copy and paste job for the rest of the new ngrx
chains. instead, got caught up debugging admin-console.
which is no longer showing family or maker.
havent figured out what is wrong, so its time to
put the vscode debugger to work!