Debugged the familyChoices returning undefined. we werent saving the return value of the map function in the variable.
Then started defining the remaing selectors, didnt
get far, before we defined a new selector in forum
and game info which selects a certain title and filters
with substring, so that we can search for a paticular family
in admin-dashboard-form.

Didnt get around to putting these into use on the UI,
want to finish implementing the selectors first.