Finished the Thread selector, which was my sole intent when begining this ngrx bulk coding task.
you can now, by using a selector, filter game info,
forum info, threads and consoles by using
title substring and the family (type) for each object.

Began laying the infastructure for a filter/ search bar
in the admin-dashboard-table-manager.
want to be careful moving forward, trying to avoid
making this a smart component, since it is supposed to
be reusable. I dont want to make this middle tier
component communicate with the store, but using
events to accomplish that might also be a mess, so
this will need some delibration.

currently, our filter in admin-dashboard-table-manager
simply obtains a query term, and tells us
what dataType (game info, forum info, thread or console)
and what family (depends on the datatype) the
query term is searching against, and this is logged
to the console. Everything is working as expected,
so we just need to figure the above  mentioned
architecture issue before implementing a UI
with this filter and moving on to admin-thread and
admin-user. lots of potential progress this week,
lets get it.