Solved our packing problem, we were misusing angularfire2, importing something that no longer exists.
our auth testing site in app component is bug free
in ide and runtime, but the logged in user is not
displaying, and we get no kind of indication that
someone is logged in/ no logout button visisble.

Im going to double back to hopefully get this
working today- both via google and email/password,
later, but for now, we commented out that simplified
auth example in app component, to pick at bugs elsewhere

admin-thread works a little better, used a bit of hard coding
to get the dashboard-form to hide whenever we take an
action with the form. Would like to get this automated,
so that every instance of the form has this behaivor, but
perhaps it can be toggled off.

two immediate items for administering threads-
1.  would be helpful to define classes for thread and metadata
so that we can just call the constructor rather than what is
done currently- defeating
the purpose of oop like seem in admin-dashboard-form
component lines 89-101.
2, need a way to toggle forum choice from dashboard form,
but also pass it in from parents so that when we select
'create-thread' while examining the nintendo forum, it
knows we want to make a nintendo forum.

finally- is checkbox really the best choice for the
forum selection area? youll see we have an issue for making
a similar component for threads.. I think the current impl
is introducing too many unknowns and very little benefits,
better to use a mat-toggle-button or a styled select. Later
on we can worry about adminstering multiple threads or
multiple forums concurrently.