Straightened out a lot with the metadata/ constructor details with thread crud. We can now create and delete and its stored in the database.
the thread selector wont display the new thread
but I think this is an issue with that component, not
the create and update features. update changes
are reflected, but we cant see newly created threads
in the interface.

this interface isnt great any way. a very shoddy prototype
of what I intended, so hoping that revising this will
eliminate this issue. There is nothing wrong with the shape
of the data in the db, so if its not an issue with the ui then
im unclear what the problem is.