While the site was working a bit ago, we were initializing ngrx effects, reducers in feature modules incorrectly. likely meaning we will need to change alot of client facing code.
Another (frightening) thought occured to me, which is
that we might have broke the app pretty bad by
upgrading packages, typecsript etc. A lot of tutorials,
git feature requests/ bug reports are at the latest
2016. meaning an oversite, we upgraded typescript
and dependencies to a point where ngrx is no longer
supported. just a theory, worth looking into,
not the wost thing since we can just use a container, or
downgrade typescript and the packages.