Solved one issue, where we couldnt see the secret keys retrieced from service when logged to console.
We can see them now, but doing something wrong
with intializing firebase. We had to do something jenky
since you provide providers in service, but inject
services into components. so our intentions with init
firebase were to import the nesc firebase libraries,
DI the envServiceProvider and initialize the firebase
app with the firebaseConfig read from the envService
injected into that modules constructors. Convoluted,
but it basically means we have to use the old typescript
class style of declaring input and export arrays in the module
class. Not sure if this is a feasible work around. but regardless
it doesnt work at this point, time to debug, look into other
solutions.