Working on getting confirm password validation working, its been tricky. We have nested the password and confirmPassword inside a passwords form group and assigned the form control to that form group, just not positive the way I did that is the correct way,
This component isnt loggng any errors or noticable
problems, but the password mismatch isnt making
an error come up in the UI like it should.

Time to get into the debugging weeds. Also we
are experimenting with two different impl for the
that validators and not sure which is right. neither
work currently which indicates it could be a issue with
component markup or logic