Added a bit of QoL stuff to nav bar, auth service,
We set off to make create user with email and password
work as well as google login, but found a lot of foundation
laying to get done along the way:
-added auth guard with return url to redirect guests
trying to reach authorization requiring pages to
the login, and then back to what they wanted to see
-navbar displays a working log out button when
user is signed in, as well as their display name and
route to profile, and login button when no one
is logged in.

- create-account - added a "dismiss" button to the info panel
about name casing that appears when a user tries
to provide a displayName in use already.

-app-component: started figuring out what needs to
happen when authorization is successful, added
store connectivity so we can update the users table,
but not much passed that, there is quite a few
branching cases from here, based on auth.provider
and whether the user is first time sign in or not.