Weird, we determined that the create-account form being invalid has nothing to do with app component, It would seem we were celebrating too soon assuming that the displayname unique validator is working perfectly (this is why you need to write tests!!!). There is some kind of async problem with the form, it is error free/ should be error free, there is no problem with the input. has status PENDING and wont submit, also display name required error appears, even when display name is provided, thats weird.
Very confused how this stopped working out of the blue
after being perfectly fine for days.

Added a user-selected service, which is technically a generic
behaivor subject that takes any observable and passes it
up the component tree, if it works that will be cool and
we can do a lot of refactoring of existing code to use a
single generic behaivor subject, plus it will make our lives
a lot easier to deal with the contributions and other
authenticated user interactions.