This thing wont budge. Time to consider other opts can we pass in a store slice to the validator argument? what if we get a list of user display names first and pass that?
We got a toy impl to work by querying firestore db
and getting matching usernames. it works but doesnt
return an observable like we need async validator to.
It also has to match the display name in database EXACTLY
so not really a practical solution more a Proof of Concept