Solved the displayNameUnique validator with a cheap non async trick. moving on but made a todo that it needs improvement.
instantly ran into another problem which involves
rxjs. So tired of this. maybe time to work on another
project?

trying to make a method for selecting the current logged in
user from entity and store based on auth state,
and have that live inside authService. I wrote it,
but currently does not work, yey (blows kazooo)