Made some minor adjustments to profile and user component, user routes- now any sub route of users/profiles has to have a route param to point to the user profile in question.
Considering back pedaling on a solution we used for
viewing games and forums, before reproducing this
implementation in other parts of the site. concerned
I may be re inventing the wheel, need to review ecmascript
docs.

Also considering whether we want profile and preferences
to have their own store slice or just make it a property
of user document with a complex object as value. This
is probably easier than reproducing existing ngrx code
for another feature.