Added a filter for the user browser, so users can search for a paticular user.
Updated user model to include profile and preferences,
this isnt finished as we need to change auth service
and ngrx code to make this property null when
storing user in db.

Made some elementary details for user-view (profile).
you can see their nickname, bio, avatar and background
image (if these are defined) when navigating to profile.
Would like to clean up this code but the more intelligent
implementation failed so we likely need to do some small
debugging before moving on with this paticular task.