Prettied up profile and did some fleshing out. you can now toggle between profile-info, preferences contributions and social links on the same component.
if the current authenticated user is the same as
the owner of profile, buttons become visible for
editing the profile, preferences and viewing
the contributions dashboard. There some kind
of finnicy problem with routing that we cant figure
out now. possible that our user auth guard is in the way,
not having a lot of luck with debugging this now.