Added some resiliency to global error handler so it doesnt cause error if the error object its attempting to handle doesnt have a message.
users.selector and its invoking services, auth guards,
validators etc- selectUserByExactDisplayName now
returns only the user object associated with that account.

create-account component- trying to trouble shoot
the displayName uniqueness validator and/or form control
for displayName. as it stands, the display name taken
error stays in markup whether or not the error is currently
in effect. For example, you provide a display name that
is taken and get the error, and delete a chunk of it,
meaning there is no longer a match, but the error
stays there regardless.

navbar- little progress here, we took the console log
we were finally getting last night and used it as a spring
board for changes made today- ideally we will finish what
were doing in auth service then come back to this and try
using a different selector such as selectById or
selectBy DispalyNameExactMatch.

user.model- ties into auth service refactoring, we got
rid of metadata property, which was unnessecary/ a quick
cheap hack.

auth service- doing quite a bit of refactoring here, which
we are in process of working through. In addition to adding
documentation for all the methods, reduced the complexity
of the googleLogin method and got some ideas for further
improvement of all the sign in and account creation related
methods. We will finish this refactoring by giving these
methods return values to tell their callers how to handle
the results of these processes, perhaps/probably using
promises.

generally, trying to improve the quality of code in
auth service instead of putting that off and continually
having to put out fires in the component logic that depends
on this service.