Got this working under a slight inconvenience.
logic/ responsiveness is good. we just had to make
a compromise that in the markup, there is only one
element rendered in either case, of auth or non auth
user, meaning we had to wrap the profile button
and logout button in a div. But this doesnt
matter much since we are going to refactor this navbar
into a few components to make it more responsive and
reactive.