Well we made a mess of this solution. putting it to rest is what I plan to do.
realizing we are getting bogged down in the weeds
trying to get this navbar in top shape.
we just need a simple way for two things
- if user is logged in, show this somehow
-based on auth state, render login or logout

we need to get moving on from this navbar
fantasy. got the styles to a good working point,
but the implementation not in a good state
since style and markup is split across two components
and needs to communicate. plan it out better in the morning.