Moved auth user dropdown into sub module of app module, because that one has connectivity to the user store.
Fixed some minor bugs in auth user dropdown. It still needs
styling and responsiveness, but th gyst of it is working
like it used to when it was part of the dropdown proper:
-can sign in via link to auth/login
-can sign out

responsiveness: for either of these behaivors, refresh
or route/link change needs to happen for the u2d data to
be displayed in UI.

styling: just needs some tweaking. we applied the same
 styles from dropdown to auth dropdown, but now only
 the signed out state (sign in button) is visible to UI.
 There are some nuances with the markup in this component
 versus the normal dropdown, those need to be accounted for
 before things look good and work properly again.