Commented out the new preference selection section. Suggesting we do this in its own routable component within the user module.
Still having issues with this initial preference form.
Not really sure whats going on with this but there
always seems to be an issue. Some things are
very unclear about the material dialog + form
interface, not from their own documentation
but in concert with template driven forms. hoping
it wasnt a mistake to switch from reactive in this
instance.